fix: logout page design issues
