+++ src/components/ErrorPage/ErrorPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRouteError, Link } from 'react-router-dom';
 
-import './syle.css'
+import './style.css'
 import Button from '../../utils/Button/Button';
 
 function ErrorPage() {
